SI-4552 feat(ui): Отдельные виджеты
