+++ OOP Patterns/aquarium/pubspec.yaml
@@ -38,7 +38,6 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
 
-  flutter_bloc: ^8.1.3
   intl: ^0.18.1
   provider: ^6.0.5
   surf_lint_rules: ^3.0.0
